login ok

+++ js/apiManager.js
@@ -1,8 +1,44 @@
 class ApiManager {
 	constructor() {}
+	getTokenFromCookie() {
+		const match = document.cookie.match(/(^| )token=([^;]+)/);
+		return match ? match[2] : null;
+	}
+	redirection () {
+		window.location.href = '/auth.html'
+	}
+	async verifyToken(token) {
+		if(!token) {
+			window.location.href = '/auth.html'
+			return false
+		}
+		try{
+			const res = await fetch('http://localhost:3000/auth/check-token', {
+				method: 'GET',
+				headers: { 'Authorization': `Bearer ${token}` }
+			})
+			console.log("Response from server:", res.status);
+			if(!res.ok) {
+				this.redirection()
+				return false
+			}
+			return true
+		} catch (e) {
+			console.log(e);
+			this.redirection();
+			return false
+		}
+	}
 	async getAllTasks() {
 		try{
-			const res = await fetch('http://localhost:3000/todos');
+			const token = this.getTokenFromCookie();
+			const isValid = await this.verifyToken(token);
+			if(!isValid) {
+				return
+			}
+			const res = await fetch('http://localhost:3000/todos', {
+				headers: token ? { 'Authorization': `Bearer ${token}` } : {},
+			});
 			if (!res.ok) {
 				throw new Error("responce not ok, getAllTasks")
 			}
@@ -13,11 +49,17 @@ class ApiManager {
 		
 	};
 	async addNewTask(task) {
+		const token = this.getTokenFromCookie();
+		const isValid = await this.verifyToken(token);
+			if(!isValid) {
+				return
+			}
 		const res = await fetch('http://localhost:3000/todos', {
 			method: 'POST',
 			body: JSON.stringify(task),
 			headers: {
 				'Content-type': 'application/json',
+				'Authorization': token ? `Bearer ${token}` : '',
 			},
 		});
 		return await res.json();
@@ -27,8 +69,14 @@ class ApiManager {
 			if (!id) {
             throw new Error("ID is required to delete a task");
 			}
+			const token = this.getTokenFromCookie();
+			const isValid = await this.verifyToken(token);
+			if(!isValid) {
+				return
+			}
 			const res = await fetch(`http://localhost:3000/todos/${id}`, {
 			method: 'DELETE',
+			headers: token ? { 'Authorization': `Bearer ${token}` } : {},
 			});
 			return await res.json();
 		} catch (err) {
@@ -36,11 +84,17 @@ class ApiManager {
 		}
 	}
 	async patchNewTask(id, updates) {
+		const token = this.getTokenFromCookie();
+		const isValid = await this.verifyToken(token);
+			if(!isValid) {
+				return
+			}
 		const res = await fetch(`http://localhost:3000/todos/${id}`, {
 			method: 'PATCH',
 			body: JSON.stringify(updates),
 			headers: {
 				'Content-Type': 'application/json',
+				'Authorization': token ? `Bearer ${token}` : '',
 			},
 		});
